[chore] Create dedicated signal routing processors.
